% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litterfluxes_AWEN.r
\name{barley.AWEN}
\alias{barley.AWEN}
\title{Calculates the AWEN proportions of barley biomass}
\usage{
barley.AWEN(Lbarley)
}
\arguments{
\item{Lbarley}{barley litter biomass}
}
\description{
Calculates the AWEN proportions of barley biomass
}
\author{
Boris Tupek
}
