% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litter_allo_AWEN.r
\name{branches.AWEN}
\alias{branches.AWEN}
\title{Calculates the AWEN proportions of branches biomass}
\usage{
branches.AWEN(Lb)
}
\arguments{
\item{Lb}{branches biomass}
}
\description{
Calculates the AWEN proportions of branches biomass
}
\seealso{
\link{branches.litter}
}
