% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litter_allo_AWEN.r
\name{dwarfshrub.litter}
\alias{dwarfshrub.litter}
\title{Calculates the biomass of the understorey dwarf shrubs}
\usage{
dwarfshrub.litter(mass, comp)
}
\description{
Calculates the biomass of the understorey dwarf shrubs
}
