% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litter_allo_AWEN.r
\name{fineroot.AWEN}
\alias{fineroot.AWEN}
\title{Calculates the AWEN proportins of fine roots biomass}
\usage{
fineroot.AWEN(Lfr, spec)
}
\arguments{
\item{Lfr}{fine roots biomass}

\item{spec}{tree species}
}
\description{
Calculates the AWEN proportins of fine roots biomass
}
