% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biomass_functions.r
\name{fineroot.linear}
\alias{fineroot.linear}
\title{Fine root based on root mass}
\usage{
fineroot.linear(Mf, spec)
}
\arguments{
\item{Mf}{mass of foliage (any mass unit)}

\item{spec}{tree species, 1 = Pine, 2 = Spruce, 3 = Birch}
}
\description{
Fine root based on root mass
}
\references{
?????????? not sure at all!!!  Helmisaari, H.-S., Derome, J., Nöjd, P. & Kukkola, M. Fine root biomass in relation to site and stand characteristics in Norway spruce and Scots pine stands. Tree Physiol. 27, 1493–1504 (2007).
}
