% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litterfluxes_AWEN.r
\name{fineroot.litter.reg}
\alias{fineroot.litter.reg}
\title{Calculates fine root litter}
\usage{
fineroot.litter.reg(Mfr, reg)
}
\arguments{
\item{Mfr}{mass of fine roots???}

\item{reg}{Region (1= Northern Finland or 2= Southern Finland)}
}
\description{
Calculates fine root litter
}
