% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litter_allo_AWEN.r
\name{fineroot.litter.tsum}
\alias{fineroot.litter.tsum}
\title{Calculates fine root litter with three different options based on tsum}
\usage{
fineroot.litter.tsum(Mfr, tsum)
}
\arguments{
\item{tsum}{temperature sum}
}
\description{
Calculates fine root litter with three different options based on tsum
}
