% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biomass_functions.r
\name{fineroot.stand}
\alias{fineroot.stand}
\title{Fine root biomass function (basal area)}
\usage{
fineroot.stand(basal, decid, CNo)
}
\arguments{
\item{basal}{stand basal area (\eqn{m^2 ha^{-1}})}

\item{decid}{dominance of birch, factor (0= dominance of conifer, 1 = dominance of birch????)}

\item{CNo}{carbon:nitrogen ratio of organic layer or upper 0–20 cm peat layer}
}
\value{
fine root biomass (<2mm) in g
}
\description{
Part of a series of functions predicting the fine root biomass of a stand
}
\references{
Lehtonen, A. et al. Modelling fine root biomass of boreal tree stands using site and stand variables. For. Ecol. Manag. 359, 361–369 (2016).
}
