% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biomass_functions.r
\name{fineroot.stand.volume.OLD}
\alias{fineroot.stand.volume.OLD}
\title{Fine root biomass function, DEPRECATED (volume)}
\usage{
fineroot.stand.volume.OLD(vol)
}
\arguments{
\item{vol}{stand volume (\eqn{m^3 ha^{-1}})}
}
\value{
fine root biomass (<2mm) in g
}
\description{
Part of a series of functions predicting the fine root biomass of a stand
}
