% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biomass_functions.r
\name{fineroot.stand.volume}
\alias{fineroot.stand.volume}
\title{Fine root biomass function (volume)}
\usage{
fineroot.stand.volume(vol)
}
\arguments{
\item{vol}{stand volume (\eqn{m^3 ha^{-1}})}
}
\value{
fine root biomass (<2mm) in g
}
\description{
Part of a series of functions predicting the fine root biomass of a stand
}
\references{
Lehtonen, A. et al. Modelling fine root biomass of boreal tree stands using site and stand variables. For. Ecol. Manag. 359, 361–369 (2016).
}
