% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biomass_functions.r
\name{fineroot.total.tsum}
\alias{fineroot.total.tsum}
\title{Fine root of the whole stand (including understorey), based on tsum}
\usage{
fineroot.total.tsum(tsum)
}
\arguments{
\item{tsum}{average annual effective temperature sum (>5 °C)}
}
\description{
Fine root of the whole stand (including understorey), based on tsum
}
\references{
?????????? not sure at all!!!  Helmisaari, H.-S., Derome, J., Nöjd, P. & Kukkola, M. Fine root biomass in relation to site and stand characteristics in Norway spruce and Scots pine stands. Tree Physiol. 27, 1493–1504 (2007).
}
