% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biomass_functions.r
\name{fineroots}
\alias{fineroots}
\title{Fine root based on root mass}
\usage{
fineroots(basal, decid, cno)
}
\arguments{
\item{Mf}{mass of roots}

\item{spec}{tree species, 1 = Pine, 2 = Spruce, 3 = Birch}
}
\description{
Fine root based on root mass
}
\references{
Helmisaari, H.-S., Derome, J., Nöjd, P. & Kukkola, M. Fine root biomass in relation to site and stand characteristics in Norway spruce and Scots pine stands. Tree Physiol. 27, 1493–1504 (2007).
}
