% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biomass_functions.r
\name{finerootsbytype}
\alias{finerootsbytype}
\title{Fine root based on root mass and type}
\usage{
finerootsbytype(Mf, spec, type)
}
\arguments{
\item{Mf}{mass of roots}

\item{spec}{tree species, 1 = Pine, 2 = Spruce, 3 = Birch}

\item{type}{Forest type, it is a proxy for soil fertility. 3 = Myrtillus, 4 = Vaccinium, 5 = Calluna}
}
\description{
Fine root based on root mass and type
}
\references{
Helmisaari, H.-S., Derome, J., Nöjd, P. & Kukkola, M. Fine root biomass in relation to site and stand characteristics in Norway spruce and Scots pine stands. Tree Physiol. 27, 1493–1504 (2007).
}
