% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litterfluxes_AWEN.r
\name{foliage.litter.Liski2006}
\alias{foliage.litter.Liski2006}
\title{Calculates foliage litter based on the older approach, with one less option (min). Deprecated.}
\usage{
foliage.litter.Liski2006(Mf, spec, reg, min)
}
\arguments{
\item{Mf}{mass of living foliage (unitless, usually in Mg ha$^{-1}$)}

\item{spec}{tree species, 1=pine 2=spruce, 3=residues (broadleaves)}

\item{reg}{Region (1= Northern Finland or 2= Southern Finland)}
}
\description{
Calculates foliage litter based on the older approach, with one less option (min). Deprecated.
}
\references{
Liski, J. et al. Carbon accumulation in Finland’s forests 1922–2004 – an estimate obtained by combination of forest inventory data with modelling of biomass, litter and soil. Ann. For. Sci. 63, 687–697 (2006).
}
\seealso{
\link{foliage.AWEN}
}
