% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litter_allo_AWEN.r
\name{foliage.litter}
\alias{foliage.litter}
\title{Calculates foliage litter}
\usage{
foliage.litter(Mf, spec, reg, min)
}
\arguments{
\item{spec}{tree species}
}
\description{
Calculates foliage litter
}
