% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litter_allo_AWEN.r
\name{lichen.AWEN}
\alias{lichen.AWEN}
\title{Calculates the AWEN proportions of lichens biomass}
\usage{
lichen.AWEN(Ll)
}
\arguments{
\item{Ll}{grass biomass}
}
\description{
Calculates the AWEN proportions of lichens biomass
}
