% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biomass_functions.r
\name{marklund.birch.livbranch}
\alias{marklund.birch.livbranch}
\title{Marklund function for birch, branches}
\usage{
marklund.birch.livbranch(d13rm)
}
\arguments{
\item{d13rm}{tree diameter at breast heigh, cm (the function converts it to dk=2+1.25*d13rm)}
}
\value{
The biomass component defined by the function (kg)
}
\description{
Marklund function for birch, branches
}
\references{
Marklund L.G., Biomassafunktioner för tall, gran och björk i Sve-rige, Sveriges lantbruksuniversitet, Rapporter-Skog 45 (1988) 1–73
}
