% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biomass_functions.r
\name{marklund.spruce.needles.comp}
\alias{marklund.spruce.needles.comp}
\title{Marklund function for spruce, needles}
\usage{
marklund.spruce.needles.comp(d13rm, ht, hlc)
}
\arguments{
\item{d13rm}{tree diameter at breast heigh, cm (the function converts it to dk=2+1.25*d13rm)}

\item{ht}{tree height, m}

\item{hlc}{?????????????}
}
\value{
The biomass component defined by the function, kg
}
\description{
Marklund function for spruce, needles
}
\references{
Marklund L.G., Biomassafunktioner för tall, gran och björk i Sve-rige, Sveriges lantbruksuniversitet, Rapporter-Skog 45 (1988) 1–73
}
