% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litter_allo_AWEN.r
\name{moss.AWEN}
\alias{moss.AWEN}
\title{Calculates the AWEN proportions of mosses biomass}
\usage{
moss.AWEN(Lm)
}
\arguments{
\item{Lm}{grass biomass}
}
\description{
Calculates the AWEN proportions of mosses biomass
}
