% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_species_AWEN.R
\name{multispecies.repo.AWEN}
\alias{multispecies.repo.AWEN}
\title{Calculates the AWEN proportions of different component based on the data in Repo et al., 2014. The function is a collection of factors and can be used for both coarse and fine components by
selecting the specific option}
\usage{
multispecies.repo.AWEN(biom, comp, type)
}
\arguments{
\item{biom}{biomass to be converted into AWEN pool, unit does not matter (output will be in the same unit)}

\item{comp}{components. "coarse" = stem, branch, roots, stump; "fine" = foliage, fine roots}

\item{type}{tree type, "temp.broad.ev" = temperate broadleaved evergreen, "temp.broad.sum" =  = temperate broadleaved evergreen, bor.broad.sum = boreal broadleaved summergreen, "temp" = temperate, "bor.con" = boreal coniferous}
}
\value{
the mass of the four AWEN components
}
\description{
Calculates the AWEN proportions of different component based on the data in Repo et al., 2014. The function is a collection of factors and can be used for both coarse and fine components by
selecting the specific option
}
\references{
Repo, A., Böttcher, H., Kindermann, G., & Liski, J. (2014). sustainability of forest bioenergy in europe: land‐use‐related carbon dioxide emissions of forest harvest residues. GCB bioenergy, 7(4), 877-887. https://doi.org/10.1111/gcbb.12179
}
\seealso{
\link{foliage.AWEN, fineroot.AWEN, branches.AWEN}
}
\author{
Lorenzo Menichetti
}
