% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biomass_functions.r
\name{petersson.spruce.below}
\alias{petersson.spruce.below}
\title{Petersson function for spruce}
\usage{
petersson.spruce.below(d13rm)
}
\arguments{
\item{d13rm}{tree diameter at breast heigh, cm (the function converts it to dk=2+1.25*d13rm)}
}
\value{
thbe belowground biomass (>2mm) in g
}
\description{
The function is an alternative approach to Repola, and it calculated the below-ground biomass to roots >2 mm. Please note that the unit is different from repola (g instead of kg).
}
\references{
Petersson, H. & Ståhl, G. Functions for below-ground biomass of Pinus sylvestris, Picea abies, Betula pendula and Betula pubescens in Sweden. Scand. J. For. Res. 21, 84–93 (2006).
}
