% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biomass_functions.r
\name{repola.pine.stump}
\alias{repola.pine.stump}
\title{Repola function for pine, stump biomass, model 1}
\usage{
repola.pine.stump(d13rm)
}
\arguments{
\item{d13rm}{tree diameter at breast heigh, cm (the function converts it to dk=2+1.25*d13rm)}
}
\value{
The biomass component defined by the function (kg)
}
\description{
Repola function for pine, stump biomass, model 1
}
\references{
Repola, J. Biomass equations for birch in Finland. Silva Fenn. 42, (2008) and Repola, J. Biomass equations for Scots pine and Norway spruce in Finland. Silva Fenn. 43, (2009).
}
