% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biomass_functions.r
\name{repola.spruce.dens}
\alias{repola.spruce.dens}
\title{Repola function for wood density, spruce}
\usage{
repola.spruce.dens(d13rm, t13)
}
\arguments{
\item{d13rm}{tree diameter at breast heigh, (cm). The function converts it to dk=2+1.25*d13rm.}

\item{t13}{tree age at breast height (years)}
}
\value{
bulk density of wood (dimensionless, ratio)
}
\description{
Repola function for wood density, spruce
}
\references{
Repola, J. & Kukkola, R. O. and M. Biomass functions for Scots pine, Norway spruce and birch in Finland. Working Papers of the Finnish Forest Research Institute (2007).
}
