% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biomass_functions.r
\name{repola.spruce.stem.simp}
\alias{repola.spruce.stem.simp}
\title{Repola function for spruce, stem wood, model 1}
\usage{
repola.spruce.stem.simp(d13rm, ht)
}
\arguments{
\item{d13rm}{tree diameter at breast heigh, cm (the function converts it to dk=2+1.25*d13rm)}

\item{ht}{tree height, m}
}
\value{
The biomass component defined by the function, kg
}
\description{
Repola function for spruce, stem wood, model 1
}
\references{
Repola, J. Biomass equations for birch in Finland. Silva Fenn. 42, (2008) and Repola, J. Biomass equations for Scots pine and Norway spruce in Finland. Silva Fenn. 43, (2009).
}
