% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litter_allo_AWEN.r
\name{root.litter}
\alias{root.litter}
\title{Calculates root litter}
\usage{
root.litter(Mr, spec)
}
\arguments{
\item{Mr}{Mass roots (unitless, usually in Mg ha$^{-1}$)}

\item{spec}{tree species, 1=pine 2=spruce, 3=residues (broadleaves)}
}
\description{
Calculates root litter
}
