% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litterfluxes_AWEN.r
\name{shit.AWEN}
\alias{shit.AWEN}
\title{Calculates the AWEN proportions of cow dung biomass}
\usage{
shit.AWEN(Lshit)
}
\arguments{
\item{Lshit}{cow dung biomass}
}
\description{
Calculates the AWEN proportions of cow dung biomass
}
\author{
Boris Tupek
}
