% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litterfluxes_AWEN.r
\name{stem.AWEN}
\alias{stem.AWEN}
\title{Calculates the AWEN proportions of stems biomass}
\usage{
stem.AWEN(Lst, spec)
}
\arguments{
\item{Lst}{branches biomass}

\item{spec}{tree species tree species}
}
\value{
the mass of the four AWEN components
}
\description{
Calculates the AWEN proportions of stems biomass
}
\references{
Yasso07 user-interface manual. J Liski, M Tuomi, J Rasinmäki - , Helsinki, 2009
}
\seealso{
\link{foliage.litter}
}
\author{
Boris Tupek
}
