% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biomass_functions.r
\name{stem.vol.birch}
\alias{stem.vol.birch}
\title{Function for calculating the stem volume, birch}
\usage{
stem.vol.birch(d13rm, ht)
}
\arguments{
\item{d13rm}{tree diameter at breast heigh, cm}

\item{ht}{tree height, m}
}
\value{
the volme of stem (\eqn{m^3})
}
\description{
Function for calculating the stem volume, birch
}
\references{
Laasasenaho, J. Taper curve and volume functions for pine, spruce and birch. at (1982).
}
