% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biomass_functions.r
\name{stem.vol.spruce.comp}
\alias{stem.vol.spruce.comp}
\title{Function for calculating the stem volume, spruce}
\usage{
stem.vol.spruce.comp(d13rm, ht, d6)
}
\arguments{
\item{d13rm}{tree diameter at breast heigh, cm}

\item{ht}{tree height, m}

\item{d6}{diameter at 6 m height (cm)}
}
\value{
the volme of stem (m$^3$)
}
\description{
Function for calculating the stem volume, spruce
}
\references{
Laasasenaho, J. Taper curve and volume functions for pine, spruce and birch. at (1982).
}
