% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biomass_functions.r
\name{stem.vol.spruce}
\alias{stem.vol.spruce}
\title{Function for calculating the stem volume, spruce}
\usage{
stem.vol.spruce(d13rm, ht)
}
\arguments{
\item{d13rm}{tree diameter at breast heigh, cm}

\item{ht}{tree height, m}
}
\value{
the volme of stem (m$^3$)
}
\description{
Function for calculating the stem volume, spruce
}
\references{
Laasasenaho, J. Taper curve and volume functions for pine, spruce and birch. at (1982).
}
