% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litterfluxes_AWEN.r
\name{twig.AWEN}
\alias{twig.AWEN}
\title{Calculates the AWEN proportions of twig biomass.}
\usage{
twig.AWEN(Lt)
}
\value{
the mass of the four AWEN components
}
\description{
DEPRECATED, please use \link{foliage.AWEN}
}
\references{
Yasso07 user-interface manual. J Liski, M Tuomi, J Rasinmäki - , Helsinki, 2009
}
\seealso{
\link{foliage.litter}
}
