% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litterfluxes_AWEN.r
\name{understorey.biomass}
\alias{understorey.biomass}
\title{Calculates the biomass of the understorey based on % cover, uplands soils}
\usage{
understorey.biomass(p_lich, p_bryoph, p_shrubs, p_grasses)
}
\arguments{
\item{spec}{tree species, 1=pine 2=spruce, 3=residues (broadleaves). For boradleaves the functions are missing, it relies on the same functions than in pine forests}
}
\description{
Calculates the biomass of the understorey based on % cover, uplands soils
}
\references{
Muukkonen, P. et al. Relationship between biomass and percentage cover in understorey vegetation of boreal coniferous forests. Silva Fenn. 40, (2006).
}
