% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litterfluxes_AWEN.r
\name{understorey.bryoph.AWEN}
\alias{understorey.bryoph.AWEN}
\title{Calculates the AWEN proportions of bryophytes (was: mosses) biomass}
\usage{
understorey.bryoph.AWEN(Lm)
}
\arguments{
\item{Lm}{moss litter biomass}
}
\value{
the mass of the four AWEN components
}
\description{
Calculates the AWEN proportions of bryophytes (was: mosses) biomass
}
\references{
Yasso07 user-interface manual. J Liski, M Tuomi, J Rasinmäki - , Helsinki, 2009
}
\seealso{
\link{understorey.bryoph.litter}
}
\author{
Boris Tupek
}
