% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litterfluxes_AWEN.r
\name{understorey.bryoph.litter}
\alias{understorey.bryoph.litter}
\title{Calculates the biomass of the understorey bryophytes}
\usage{
understorey.bryoph.litter(mass)
}
\description{
Calculates the biomass of the understorey bryophytes
}
\seealso{
\link{understorey.bryoph}
}
