% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biomass_functions.r
\name{understorey.bryoph.OLD}
\alias{understorey.bryoph.OLD}
\title{Understorey vegetation functions, developed by Lehtonen based on data from Salemaa, M., et. al 2013}
\usage{
understorey.bryoph.OLD(cov, reg)
}
\arguments{
\item{cov}{coverage, %}

\item{reg}{region (1 = south 2 = north Finland ????)}
}
\description{
Understorey vegetation functions, developed by Lehtonen based on data from Salemaa, M., et. al 2013
}
