% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litterfluxes_AWEN.r
\name{understorey.dwarfshrub.litter}
\alias{understorey.dwarfshrub.litter}
\title{Calculates the biomass of the understorey dwarf shrubs}
\usage{
understorey.dwarfshrub.litter(mass, comp)
}
\arguments{
\item{mass}{mass of shrubs}

\item{comp}{compartment, abv = above, bel = below}
}
\value{
biomass of dwarf shrubs
}
\description{
Calculates the biomass of the understorey dwarf shrubs
}
\seealso{
\code{\link{understorey.dwarfshrub}}
}
