% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biomass_functions.r
\name{understorey.dwarfshrub.OLD}
\alias{understorey.dwarfshrub.OLD}
\title{Understorey vegetation functions, probably developed by Lehtonen, origin unknown}
\usage{
understorey.dwarfshrub.OLD(cov, comp, reg)
}
\arguments{
\item{cov}{coverage, % ???? it's probably not % but 0 to 1 ratio!!!!}

\item{comp}{compartment, abv = above, bel = below ????}

\item{reg}{region (1 = south 2 = north Finland ????)}
}
\description{
Understorey vegetation functions, probably developed by Lehtonen, origin unknown
}
