% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biomass_functions.r
\name{understorey.dwarfshrub}
\alias{understorey.dwarfshrub}
\title{Understorey vegetation functions, developed by Lehtonen based on data from Salemaa, M., et. al 2013}
\usage{
understorey.dwarfshrub(cov, comp, reg)
}
\arguments{
\item{cov}{coverage, as 0 to 1 ratio}

\item{comp}{compartment, abv = above, bel = below ????}

\item{reg}{region (1 = south 2 = north Finland ????)}
}
\description{
Understorey vegetation functions, developed by Lehtonen based on data from Salemaa, M., et. al 2013
}
\references{
?????? probably missing, data source could be Salemaa, M., Hamberg, L., Kalinauskaite, N., Korpela, L., Lindroos, A-J., Nöjd, P., & Tonteri, T. (2013). Understorey vegetation on level II plots during 1998-2009. In L. Merilä, & S. Jortikka (Eds.), Forest Condition Monitoring in Finland - National Report. The Finnish Forest Research Institute The Finnish Forest Research Institute. http://urn.fi/URN:NBN:fi:metla-201305087568
}
