% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litter_allo_AWEN.r
\name{understorey.grass.AWEN}
\alias{understorey.grass.AWEN}
\title{Calculates the AWEN proportions of grass biomass}
\usage{
understorey.grass.AWEN(Lg, above)
}
\arguments{
\item{Lg}{grass biomass}
}
\description{
Calculates the AWEN proportions of grass biomass
}
\seealso{
\link{understorey.grass.litter}
}
