% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biomass_functions.r
\name{understorey.grass}
\alias{understorey.grass}
\title{Understorey vegetation functions, developed by Lehtonen based on data from Salemaa, M., et. al 2013}
\usage{
understorey.grass(cov, comp, reg)
}
\arguments{
\item{cov}{coverage, % ???? it's probably not % but 0 to 1 ratio!!!!}

\item{comp}{compartment, abv = above, bel = below ????}

\item{reg}{region (1 = south 2 = north Finland ????)}
}
\description{
Understorey vegetation functions, developed by Lehtonen based on data from Salemaa, M., et. al 2013
}
