% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litter_allo_AWEN.r
\name{understorey.herb.litter}
\alias{understorey.herb.litter}
\title{Calculates the biomass of the understorey herbs}
\usage{
understorey.herb.litter(mass, comp)
}
\description{
Calculates the biomass of the understorey herbs
}
\seealso{
\link{understorey.herb}
}
