% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biomass_functions.r
\name{understorey.herb.OLD}
\alias{understorey.herb.OLD}
\title{Understorey vegetation functions, developed by Lehtonen based on data from Salemaa, M., et. al 2013}
\usage{
understorey.herb.OLD(cov, comp, reg)
}
\arguments{
\item{cov}{coverage, %}

\item{comp}{compartment, abv = above, bel = below ????}

\item{reg}{region (1 = south 2 = north Finland ????)}
}
\description{
Understorey vegetation functions, developed by Lehtonen based on data from Salemaa, M., et. al 2013
}
