% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biomass_functions.r
\name{understorey.lich.OLD}
\alias{understorey.lich.OLD}
\title{Understorey vegetation functions, developed by Lehtonen based on data from Salemaa, M., et. al 2013}
\usage{
understorey.lich.OLD(cov)
}
\arguments{
\item{cov}{coverage ????}
}
\description{
Understorey vegetation functions, developed by Lehtonen based on data from Salemaa, M., et. al 2013
}
