% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litter_allo_AWEN.r
\name{understorey.lichen.AWEN}
\alias{understorey.lichen.AWEN}
\title{Calculates the AWEN proportions of lichens biomass}
\usage{
understorey.lichen.AWEN(Ll)
}
\arguments{
\item{Ll}{lichens biomass}
}
\description{
Calculates the AWEN proportions of lichens biomass
}
\seealso{
\link{understorey.lichen.litter}
}
