% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biomass_functions.r
\name{understorey.lichen}
\alias{understorey.lichen}
\title{Understorey vegetation functions, developed by Lehtonen based on data from Salemaa, M., et. al 2013}
\usage{
understorey.lichen(cov)
}
\arguments{
\item{cov}{coverage, % ???? it's probably not % but 0 to 1 ratio!!!!}
}
\description{
Understorey vegetation functions, developed by Lehtonen based on data from Salemaa, M., et. al 2013
}
