% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litter_allo_AWEN.r
\name{wheat.AWEN}
\alias{wheat.AWEN}
\title{Calculates the AWEN proportions of wheat biomass}
\usage{
wheat.AWEN(Lwheat)
}
\arguments{
\item{Lwheat}{wheat biomass}
}
\description{
Calculates the AWEN proportions of wheat biomass
}
