% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Yasso_model_functions.R
\name{Yasso07.Modelfi}
\alias{Yasso07.Modelfi}
\title{Yasso07 model as in Tuomi et al. 2011 ecological applications.}
\usage{
Yasso07.Modelfi(
  t,
  ksY = c(kA = 0.73, kW = 5.8, kE = 0.29, kN = 0.031, kH = 0.0017),
  pY = c(p1 = 0.48, p2 = 0.01, p3 = 0.83, p4 = 0.99, p5 = 0, p6 = 0.01, p7 = 0, p8 = 0,
    p9 = 0.03, p10 = 0, p11 = 0.01, p12 = 0.92, pH = 0.0045),
  beta1 = 0.096,
  beta2 = -0.0014,
  gamma = -1.21,
  delta1 = -1.7,
  delta2 = 0.86,
  r = -0.306,
  C0,
  In,
  AWEN,
  xi = T,
  xi_modifier = 1,
  MT = NULL,
  TA = NULL,
  PR_mm = NULL,
  WS,
  solver = deSolve.lsoda.wrapper,
  pass = FALSE
)
}
\arguments{
\item{t}{time (years)}

\item{ksY}{a vector with the decomposition constants of the five Yasso decomposing pools, default c(kA = 0.73, kW = 5.8, kE = 0.29, kN = 0.031, kH = 0.0017)}

\item{pY}{a vector with the transfers (including feedbacks) between the five
Yasso pools, defailt  pY = c(p1 = 0.48, p2 = 0.01, p3 = 0.83, p4 = 0.99, p5
= 0, p6 = 0.01, p7 = 0, p8 = 0, p9 = 0.03, p10 = 0, p11 = 0.01, p12 = 0.92,
pH = 0.0045),}

\item{beta1}{temperature dependence parameter}

\item{beta2}{temperature dependence parameter}

\item{delta1}{Woody litter size dependence parameters (cm^-1)}

\item{delta2}{Woody litter size dependence parameters (cm^-1)}

\item{r}{Woody litter size dependence parameters (adimensional)}

\item{C0}{initial C , 5 element vector, e.g. c(0,0,0,0,0)}

\item{In}{litter C input, same length as years, if AWEN fractionation not provided it has to be a data.frame (years, 5 AWEN pools), as long as the time steps}

\item{AWEN}{5 element vector, fractionation of plant C litter input to yasso AWEN pools}

\item{xi}{climate scaling, if xi = FALSE then the model will scale decomposition based on temperature and precipitation data, otherwise it will ignore it (xi = TRUE or xi = 1 is equivalent to setting the xi scaling factor to 1)}

\item{xi_modifier}{linear modifier of the climatic effect on decomposition, if = 1 effect is null. Used mostly for model calibration.}

\item{MT}{Annual mean temperature (vector of length t) (cat('\u00B0') C)}

\item{TA}{Temperature Amplitude = (mothly temp. range)/2  (cat('\u00B0') C) (vector of length t)}

\item{PR_mm, }{Annual precipitation_mm (vector of length t)}

\item{WS}{woody size, for example 0 no effect  for nonwoody, 2 finewoody, 20 coarse woody (cm)}

\item{solver}{the solver used by deSolve to solve the linear ODE system, called by SoiLR, default is = deSolve.lsoda.wrapper,}

\item{pass}{= FALSE}

\item{gammma}{precipitation dependence parameter}
}
\description{
The function relies on the soilR implementation, and it generates a model object that needs then to be solved by soilR functions (see examples)
}
\examples{
years=seq(1:10)
Litter= cbind(years, matrix(rep(c(2,2,2,2, 0), each = 10), nrow = 10))
PR_mm=rep(200, 10)
MT= rep(10,10)
TA= rep(2,10)

yassofix <- Yasso07Modelfi(years,
                           C0=rep(0,5), #initial carbon
                           AWEN = c(0.52,0.18,0.08,0.2,0), #to separate litter to yasso AWEN pools, this depends on plant organ and species
                           In=Litter,#litter C input (same length as years)
                           xi = 1, # only xi = 1  will replace climate data no climate effect,
                           MT=MT,#MeanTemperature
                           TA=TA, #TemperatureAmplitude
                           PR_mm=PR_mm,#Precipitation_mm)
                           WS=2)

Ct=getC(yassofix)
Rt=getReleaseFlux(yassofix) #respiration

}
\references{
Ťupek, B. et al. Modeling boreal forest&rsquo;s mineral soil and peat C stock dynamics with Yasso07 model coupled with updated moisture modifier. EGUsphere 2023, 1–34 (2023).
}
\seealso{
\link{Yasso07Modelfi.month}
}
\author{
Boris Tupek, boris.tupek@luke.fi
}
