% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Yasso_model_functions.R
\name{yasso07.SS}
\alias{yasso07.SS}
\title{Yasso07 steady state calculation}
\usage{
yasso07.SS(
  WS,
  clim,
  wetlands = "n",
  In_ave,
  ksY = c(kA = 0.73, kW = 5.8, kE = 0.29, kN = 0.031, kH = 0.0017),
  pY = c(p1 = 0.48, p2 = 0.01, p3 = 0.83, p4 = 0.99, p5 = 0, p6 = 0.01, p7 = 0, p8 = 0,
    p9 = 0.03, p10 = 0, p11 = 0.01, p12 = 0.92, pH = 0.0045),
  beta1 = 0.096,
  beta2 = -0.0014,
  gamma = -1.21,
  delta1 = -1.7,
  delta2 = 0.86,
  r = -0.306
)
}
\arguments{
\item{clim}{a vector containing the mean values of MT, TA, PR_mm as in \link{Yasso07Modelfi}. If clim = 1 ignored. Since we are considering steady states, climate must be constant}

\item{wetlands}{if wetlands = "y", apply 35% reduction of decomposition se Kleinen et al. (2021). default is "n"}

\item{In_ave}{average AWEN inputs}
}
\description{
The function is just a wrapper for the example in \link{yasso.matrix.fun}, to calculate the steady state by solving the matrix model definition
}
\references{
Ťupek, B. et al. Modeling boreal forest&rsquo;s mineral soil and peat C stock dynamics with Yasso07 model coupled with updated moisture modifier. EGUsphere 2023, 1–34 (2023).
}
\author{
Lorenzo Menichetti, lorenzo.menichetti@luke.fi
}
