% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Yasso_model_functions.R
\name{Yasso07Modelfi}
\alias{Yasso07Modelfi}
\title{Yasso07 model as in Tuomi et al. 2011 ecological applications}
\usage{
Yasso07Modelfi(
  t,
  ksY = c(kA = 0.73, kW = 5.8, kE = 0.29, kN = 0.031, kH = 0.0017),
  pY = c(p1 = 0.48, p2 = 0.01, p3 = 0.83, p4 = 0.99, p5 = 0, p6 = 0.01, p7 = 0, p8 = 0,
    p9 = 0.03, p10 = 0, p11 = 0.01, p12 = 0.92, pH = 0.0045),
  beta1 = 0.096,
  beta2 = -0.0014,
  gamma = -1.21,
  delta1 = -1.7,
  delta2 = 0.86,
  r = -0.306,
  C0,
  In,
  AWEN,
  xi = 0,
  MT,
  TA,
  PR_mm,
  WS,
  solver = deSolve.lsoda.wrapper,
  pass = FALSE
)
}
\arguments{
\item{t}{TO BE DONE}

\item{ksY}{TO BE DONE}

\item{pY}{TO BE DONE}

\item{beta1}{TO BE DONE}

\item{beta2}{TO BE DONE}

\item{delta1}{TO BE DONE}

\item{delta2}{TO BE DONE}

\item{r}{TO BE DONE}

\item{C0}{TO BE DONE}

\item{In}{TO BE DONE}

\item{AWEN}{TO BE DONE}

\item{xi}{TO BE DONE}

\item{MT}{MeanTemperature}

\item{TA}{TemperatureAmplitude = (mothly temp. range)/2}

\item{PR_mm, }{Precipitation_mm}

\item{WS}{woody size, 0 no effect  for nonwoody, 2 finewoody, 20 coarse woody}

\item{solver}{the solver used by deSolve to solve the linear ODE system, called by SoiLR, default is = deSolve.lsoda.wrapper,}

\item{pass}{= FALSE}

\item{gammma}{TO BE DONE}
}
\description{
Yasso07 model as in Tuomi et al. 2011 ecological applications
}
