% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litterfluxes_AWEN.r
\name{barley.AWEN}
\alias{barley.AWEN}
\title{Calculates the AWEN proportions of barley biomass}
\usage{
barley.AWEN(Lbarley)
}
\arguments{
\item{Lbarley}{barley litter biomass}
}
\value{
the mass of the four AWEN components
}
\description{
Calculates the AWEN proportions of barley biomass
}
\references{
Yasso07 user-interface manual. J Liski, M Tuomi, J Rasinmäki - , Helsinki, 2009
}
\seealso{
\link{foliage.litter}
}
\author{
Boris Tupek
}
