% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litterfluxes_AWEN.r
\name{branches.AWEN}
\alias{branches.AWEN}
\title{Calculates the AWEN proportions of branches biomass}
\usage{
branches.AWEN(Lb)
}
\arguments{
\item{Lb}{branches biomass}
}
\value{
the mass of the four AWEN components
}
\description{
Calculates the AWEN proportions of branches biomass
}
\references{
Yasso07 user-interface manual. J Liski, M Tuomi, J Rasinmäki - , Helsinki, 2009
}
\seealso{
\link{branches.litter}
}
