% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litter_allo_AWEN.r
\name{branches.litter}
\alias{branches.litter}
\title{Calculates branch litter}
\usage{
branches.litter(Mb, spec)
}
\arguments{
\item{Mb}{Mass branches (unitless, usually in Mg ha$^{-1}$)}

\item{spec}{tree species, 1=pine 2=spruce, 3=residues (broadleaves)}
}
\description{
Calculates branch litter
}
