% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litter_allo_AWEN.r
\name{carbon}
\alias{carbon}
\title{Converts biomass in carbon}
\usage{
carbon(mass)
}
\arguments{
\item{mass}{mass of organic matter}
}
\value{
a scalar, mass of C
}
\description{
The function assumes 50\% C content of organic matter
}
