% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litterfluxes_AWEN.r
\name{fineroot.AWEN}
\alias{fineroot.AWEN}
\title{Calculates the AWEN proportins of fine roots biomass}
\usage{
fineroot.AWEN(Lfr, spec)
}
\arguments{
\item{Lfr}{fine roots biomass}

\item{spec}{tree species}
}
\value{
the mass of the four AWEN components
}
\description{
Calculates the AWEN proportins of fine roots biomass
}
\references{
Yasso07 user-interface manual. J Liski, M Tuomi, J Rasinmäki - , Helsinki, 2009
}
\seealso{
\code{\link{fineroot.litter.reg}} \code{\link{fineroot.litter.tsum}}
}
\author{
Boris Tupek
}
