% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litterfluxes_AWEN.r
\name{fineroot.litter.reg}
\alias{fineroot.litter.reg}
\title{Calculates fine root litter}
\usage{
fineroot.litter.reg(Mfr, reg)
}
\arguments{
\item{Mfr}{mass of fine roots}

\item{reg}{Region (1= Northern Finland or 2= Southern Finland)}
}
\value{
litter inputs (mass) from a mass of fine roots, to be passed to the AWEN partitioning function
}
\description{
Calculates fine root litter
}
\seealso{
\code{\link{fineroot.AWEN}}
}
