% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litterfluxes_AWEN.r
\name{fineroot.litter.tsum}
\alias{fineroot.litter.tsum}
\title{Calculates fine root litter with three different options based on tsum}
\usage{
fineroot.litter.tsum(Mfr, tsum)
}
\arguments{
\item{Mfr}{mass of fine roots}

\item{tsum}{temperature sum}
}
\value{
litter inputs (mass) from a mass of fine roots, to be passed to the AWEN partitioning function
}
\description{
Calculates fine root litter with three different options based on tsum
}
\seealso{
\code{\link{fineroot.AWEN}}
}
