% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litter_allo_AWEN.r
\name{foliage.AWEN}
\alias{foliage.AWEN}
\title{Calculates the AWEN proportins of foliage biomass}
\usage{
foliage.AWEN(Lf, spec)
}
\arguments{
\item{Lf}{leaves biomass}

\item{spec}{tree species}
}
\description{
Calculates the AWEN proportins of foliage biomass
}
