% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litterfluxes_AWEN.r
\name{foliage.AWEN}
\alias{foliage.AWEN}
\title{Calculates the AWEN proportins of foliage biomass}
\usage{
foliage.AWEN(Lf, spec)
}
\arguments{
\item{Lf}{leaves biomass}

\item{spec}{tree species}
}
\value{
the mass of the four AWEN components
}
\description{
Calculates the AWEN proportins of foliage biomass
}
\references{
Yasso07 user-interface manual. J Liski, M Tuomi, J Rasinmäki - , Helsinki, 2009
}
\seealso{
\link{foliage.litter}
}
\author{
Boris Tupek
}
