% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litterfluxes_AWEN.r
\name{foliage.litter}
\alias{foliage.litter}
\title{Calculates foliage litter}
\usage{
foliage.litter(Mf, spec, reg, min)
}
\arguments{
\item{Mf}{mass of living foliage (unitless, usually in Mg ha$^{-1}$)}

\item{spec}{tree species, 1=pine 2=spruce, 3=residues (broadleaves)}

\item{reg}{Region (1= Northern Finland or 2= Southern Finland)}

\item{min}{mineral soil, TRUE/FALSE}
}
\description{
Calculates foliage litter
}
\references{
Liski, J. et al. Carbon accumulation in Finland’s forests 1922–2004 – an estimate obtained by combination of forest inventory data with modelling of biomass, litter and soil. Ann. For. Sci. 63, 687–697 (2006).
}
\seealso{
\link{foliage.AWEN}
}
