% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litter_allo_AWEN.r
\name{foliage.litter}
\alias{foliage.litter}
\title{Calculates foliage litter}
\usage{
foliage.litter(Mf, spec, reg, min)
}
\arguments{
\item{Mf}{mass of foliage (unitless, usually in Mg ha$^{-1}$)}

\item{spec}{tree species, 1=pine 2=spruce, 3=residues (broadleaves)}
}
\description{
Calculates foliage litter
}
\seealso{
\link{foliage.AWEN}
}
