% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litterfluxes_AWEN.r
\name{shit.AWEN}
\alias{shit.AWEN}
\title{Calculates the AWEN proportions of cow dung biomass}
\usage{
shit.AWEN(Lshit)
}
\arguments{
\item{Lshit}{cow dung biomass}
}
\value{
the mass of the four AWEN components
}
\description{
Calculates the AWEN proportions of cow dung biomass
}
\references{
Yasso07 user-interface manual. J Liski, M Tuomi, J Rasinmäki - , Helsinki, 2009
}
\seealso{
\link{foliage.litter}
}
\author{
Boris Tupek
}
