% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litter_allo_AWEN.r
\name{stem.AWEN}
\alias{stem.AWEN}
\title{Calculates the AWEN proportions of stems biomass}
\usage{
stem.AWEN(Lst, spec)
}
\arguments{
\item{Lst}{branches biomass}

\item{spec}{tree species tree species}
}
\description{
Calculates the AWEN proportions of stems biomass
}
