% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litter_allo_AWEN.r
\name{stump.litter}
\alias{stump.litter}
\title{Calculates stump bark litter}
\usage{
stump.litter(Mst, spec)
}
\arguments{
\item{Mst}{Mass stumps (unitless, usually in Mg ha$^{-1}$)}

\item{spec}{tree species, 1=pine 2=spruce, 3=residues (broadleaves)}
}
\description{
Calculates stump bark litter
}
