% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litter_allo_AWEN.r
\name{understorey.dwarfshrub.litter}
\alias{understorey.dwarfshrub.litter}
\title{Calculates the biomass of the understorey dwarf shrubs}
\usage{
understorey.dwarfshrub.litter(mass, comp)
}
\arguments{
\item{comp}{compartment, abv = above, bel = below ????}
}
\description{
Calculates the biomass of the understorey dwarf shrubs
}
\seealso{
\link{understorey.dwarfshrub}
}
