% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litterfluxes_AWEN.r
\name{understorey.grass.AWEN}
\alias{understorey.grass.AWEN}
\title{Calculates the AWEN proportions of grass biomass}
\usage{
understorey.grass.AWEN(Lg, above)
}
\arguments{
\item{Lg}{grass biomass}
}
\value{
the mass of the four AWEN components
}
\description{
Calculates the AWEN proportions of grass biomass
}
\references{
Yasso07 user-interface manual. J Liski, M Tuomi, J Rasinmäki - , Helsinki, 2009
}
\seealso{
\link{understorey.grass.litter}
}
