% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litterfluxes_AWEN.r
\name{understorey.grass.litter}
\alias{understorey.grass.litter}
\title{Calculates the biomass of the understorey grasses}
\usage{
understorey.grass.litter(mass, comp)
}
\description{
Calculates the biomass of the understorey grasses
}
\seealso{
\link{understorey.grass}
}
