% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litterfluxes_AWEN.r
\name{understorey.herb.litter}
\alias{understorey.herb.litter}
\title{Calculates the litter input of the understorey herbs}
\usage{
understorey.herb.litter(mass, comp)
}
\arguments{
\item{mass}{mass of herbs}

\item{comp}{compartment, abv = above, bel = below}
}
\value{
litter inputs (mass) from a mass of herbs, to be passed to the AWEN partitioning function
}
\description{
Calculates the litter input of the understorey herbs
}
\seealso{
\code{\link{understorey.herb}} \code{\link{understorey.grass.AWEN}}
}
