% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litterfluxes_AWEN.r
\name{understorey.lichen.AWEN}
\alias{understorey.lichen.AWEN}
\title{Calculates the AWEN proportions of lichens biomass}
\usage{
understorey.lichen.AWEN(Ll)
}
\arguments{
\item{Ll}{lichens litter biomass}
}
\description{
Calculates the AWEN proportions of lichens biomass
}
\seealso{
\code{\link{understorey.lichen.litter}}
}
