% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litterfluxes_AWEN.r
\name{understorey.lichen.AWEN}
\alias{understorey.lichen.AWEN}
\title{Calculates the AWEN proportions of lichens biomass}
\usage{
understorey.lichen.AWEN(Ll)
}
\arguments{
\item{Ll}{lichens litter biomass}
}
\value{
the mass of the four AWEN components
}
\description{
Calculates the AWEN proportions of lichens biomass
}
\references{
Yasso07 user-interface manual. J Liski, M Tuomi, J Rasinmäki - , Helsinki, 2009
}
\seealso{
\link{understorey.lichen.litter}
}
\author{
Boris Tupek
}
