% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litterfluxes_AWEN.r
\name{understorey.lichen.litter}
\alias{understorey.lichen.litter}
\title{Calculates the biomass of the understorey lichens}
\usage{
understorey.lichen.litter(mass)
}
\description{
Calculates the biomass of the understorey lichens
}
\seealso{
\link{understorey.lichen}
}
