% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litterfluxes_AWEN.r
\name{wheat.AWEN}
\alias{wheat.AWEN}
\title{Calculates the AWEN proportions of wheat biomass}
\usage{
wheat.AWEN(Lwheat)
}
\arguments{
\item{Lwheat}{wheat litter biomass}
}
\value{
the mass of the four AWEN components
}
\description{
Calculates the AWEN proportions of wheat biomass
}
\references{
Yasso07 user-interface manual. J Liski, M Tuomi, J Rasinmäki - , Helsinki, 2009
}
\seealso{
\link{foliage.litter}
}
\author{
Boris Tupek
}
